backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
numpy>=1.26.0
rasterio>=1.3.10
Pillow>=10.3.0
torch>=2.3.0
torchvision>=0.18.0
python-multipart>=0.0.9
httpx>=0.27.0
python-dotenv>=1.0.1
pydantic-settings>=2.2.1
